Beta 3.0 parentView removal in Todo example code
With the currently Beta 3.0 removal of parentView, I was going through
the Getting started guide, brand new to Sproutcore and pretty new to JS.
I ran into the error of not being able to apple the isDone css, and ran
across in the Beta 3.0 parentView is no longer needed in this context. I
removed it from these two lines. First time committing to a project so I
hope this is the correct way to do things.